refactor: split oversized Type.kt files into focused single-responsibility objects
- ElixirErlangClasspath      – Erlang classpath add/remove operations
- ElixirSdkPathConfigurator  – SDK path setup (dead addDocumentationPaths removed)
- ElixirVersionDetector      – Elixir version detection and caching
- ElixirBuildInfo            – BEAM file parsing (Problem 5 landing zone)
- ElixirSdkLookup            – PSI/module/project SDK accessors
- ErlangVersionDetector      – Erlang version detection and caching
- ErlangSdkTableListener     – JDK table lifecycle listener
- SdkProcessUtils            – Shared runWithEdtGuard utility
- ErlangSdkResolver          - move findRegisteredErlangSdk here

No logic changes. All existing call sites updated.
