+++ src/org/elixir_lang/espec/configuration/Producer.kt
@@ -14,7 +14,7 @@ import org.elixir_lang.file.containsFileWithSuffix
 import org.elixir_lang.mix.Project
 import org.elixir_lang.psi.ElixirFile
 import org.elixir_lang.sdk.elixir.Type
-import org.elixir_lang.sdk.elixir.Type.Companion.mostSpecificSdk
+import org.elixir_lang.sdk.elixir.ElixirSdkLookup.mostSpecificSdk
 import java.io.File
 
 class MixESpecRunConfigurationProducer :
