+++ src/org/elixir_lang/sdk/erlang_dependent/AdditionalDataConfigurable.kt
@@ -15,9 +15,9 @@ import com.intellij.openapi.util.Comparing
 import com.intellij.ui.SimpleListCellRenderer
 import com.intellij.util.ui.JBUI
 import org.elixir_lang.debug
-import org.elixir_lang.sdk.elixir.Type.Companion.addNewCodePathsFromInternErlangSdk
+import org.elixir_lang.sdk.elixir.ElixirErlangClasspath.addNewCodePathsFromInternErlangSdk
+import org.elixir_lang.sdk.elixir.ElixirErlangClasspath.removeCodePathsFromInternalErlangSdk
 import org.elixir_lang.sdk.elixir.Type.Companion.hasErlangClasspathInRoots
-import org.elixir_lang.sdk.elixir.Type.Companion.removeCodePathsFromInternalErlangSdk
 import java.awt.GridBagConstraints
 import java.awt.GridBagLayout
 import java.awt.event.ItemEvent
